Add accordion toggle functionality to FooterAccordion component
+ Change name for footer data attributes
