+++ src/stories/Library/footer-accordions/FooterAccordions.tsx
@@ -1,14 +1,10 @@
-import { useEffect } from "react";
 import FooterAccordion from "./FooterAccordion";
 import { footerContent } from "../../Blocks/footer/footer-content";
 
 const FooterAccordions = () => {
-  useEffect(() => {
-    /* eslint-disable-next-line global-require */
-    require("./initaccordion");
-  }, []);
+  // use the logic in initaccordion.js to initialize the accordion ind drupal
   return (
-    <ul className="footer-accordions" onClick={(e) => window.eventAccordion(e)}>
+    <ul className="footer-accordions" data-footer-accordions>
       {footerContent.map(({ title, content }, i) => (
         <li className="footer-accordion" key={i}>
           <FooterAccordion title={title} content={content} open={i === 0} />
@@ -17,13 +13,5 @@ const FooterAccordions = () => {
     </ul>
   );
 };
-
-declare global {
-  interface Window {
-    eventAccordion: (
-      event: React.MouseEvent<HTMLUListElement, MouseEvent>
-    ) => void;
-  }
-}
 
 export default FooterAccordions;
