--- src/stories/Library/footer-accordions/initaccordion.js
@@ -1,47 +1,0 @@
-// function accordion(event) {
-//   if (!event?.target?.closest) {
-//     // eslint-disable-next-line
-//     console.info("Accordion - No event target");
-//     return;
-//   }
-
-//   const btn = event.target.closest("[data-accordion-trigger]");
-
-//   if (!btn?.getAttribute) {
-//     // eslint-disable-next-line
-//     console.info("Accordion - Can't find accordion row");
-//     return;
-//   }
-
-//   if (!btn.parentNode?.nextElementSibling) {
-//     // eslint-disable-next-line
-//     console.info("Accordion - Can't find next sibling to parentNode");
-//     return;
-//   }
-
-//   const isOpen = btn.getAttribute("aria-expanded") === "true";
-//   const target = btn.parentNode.nextElementSibling;
-
-//   btn.setAttribute("aria-expanded", !isOpen);
-//   target.hidden = isOpen;
-// }
-
-// window.eventAccordion = (e) => {
-//   accordion(e);
-// };
-
-document.addEventListener("DOMContentLoaded", () => {
-  document
-    .querySelector(".footer-accordions")
-    .addEventListener("click", (e) => {
-      const btn = e.target.closest("[data-accordion-trigger]");
-
-      if (!btn) return;
-
-      const isOpen = btn.getAttribute("aria-expanded") === "true";
-      const target = btn.parentNode.nextElementSibling;
-
-      btn.setAttribute("aria-expanded", String(!isOpen));
-      target.hidden = isOpen;
-    });
-});
